Add files via upload
The task was to Construct a program that functions as a basic arithmetic calculator. The program shall prompt the user for two numerical inputs and an operation selection (+, -, , /). Upon receiving the user's input, the program shall perform the chosen operation on the provided numbers and subsequently display the calculated result.